Fix octoprint tests opening sockets
