+++ tests/components/octoprint/test_binary_sensor.py
@@ -1,23 +1,33 @@
 """The tests for Octoptint binary sensor module."""
 
+import pytest
+
 from homeassistant.const import STATE_OFF, STATE_ON, STATE_UNAVAILABLE
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers import entity_registry as er
 
-from . import init_integration
+
+@pytest.fixture
+def platform() -> str:
+    """Fixture to specify platform."""
+    return "binary_sensor"
 
 
+@pytest.mark.parametrize(
+    "printer",
+    [
+        {
+            "state": {
+                "flags": {"printing": True, "error": False},
+                "text": "Operational",
+            },
+            "temperature": [],
+        },
+    ],
+)
+@pytest.mark.usefixtures("init_integration")
 async def test_sensors(hass: HomeAssistant, entity_registry: er.EntityRegistry) -> None:
     """Test the underlying sensors."""
-    printer = {
-        "state": {
-            "flags": {"printing": True, "error": False},
-            "text": "Operational",
-        },
-        "temperature": [],
-    }
-    await init_integration(hass, "binary_sensor", printer=printer)
-
     state = hass.states.get("binary_sensor.octoprint_printing")
     assert state is not None
     assert state.state == STATE_ON
@@ -33,12 +43,12 @@ async def test_sensors(hass: HomeAssistant, entity_registry: er.EntityRegistry) 
     assert entry.unique_id == "Printing Error-uuid"
 
 
+@pytest.mark.parametrize("printer", [None])
+@pytest.mark.usefixtures("init_integration")
 async def test_sensors_printer_offline(
     hass: HomeAssistant, entity_registry: er.EntityRegistry
 ) -> None:
     """Test the underlying sensors when the printer is offline."""
-    await init_integration(hass, "binary_sensor", printer=None)
-
     state = hass.states.get("binary_sensor.octoprint_printing")
     assert state is not None
     assert state.state == STATE_UNAVAILABLE
