+++ tests/components/octoprint/test_button.py
@@ -2,7 +2,6 @@
 
 from unittest.mock import patch
 
-from freezegun import freeze_time
 from pyoctoprintapi import OctoprintPrinterInfo
 import pytest
 
@@ -13,13 +12,16 @@ from homeassistant.components.octoprint.const import DOMAIN
 from homeassistant.const import ATTR_ENTITY_ID
 from homeassistant.core import HomeAssistant
 
-from . import init_integration
+
+@pytest.fixture
+def platform() -> str:
+    """Fixture to specify platform."""
+    return BUTTON_DOMAIN
 
 
+@pytest.mark.usefixtures("init_integration")
 async def test_pause_job(hass: HomeAssistant) -> None:
     """Test the pause job button."""
-    await init_integration(hass, BUTTON_DOMAIN)
-
     coordinator: OctoprintDataUpdateCoordinator = hass.data[DOMAIN]["uuid"][
         "coordinator"
     ]
@@ -75,10 +77,9 @@ async def test_pause_job(hass: HomeAssistant) -> None:
             )
 
 
+@pytest.mark.usefixtures("init_integration")
 async def test_resume_job(hass: HomeAssistant) -> None:
     """Test the resume job button."""
-    await init_integration(hass, BUTTON_DOMAIN)
-
     coordinator: OctoprintDataUpdateCoordinator = hass.data[DOMAIN]["uuid"][
         "coordinator"
     ]
@@ -134,10 +135,9 @@ async def test_resume_job(hass: HomeAssistant) -> None:
             )
 
 
+@pytest.mark.usefixtures("init_integration")
 async def test_stop_job(hass: HomeAssistant) -> None:
     """Test the stop job button."""
-    await init_integration(hass, BUTTON_DOMAIN)
-
     coordinator: OctoprintDataUpdateCoordinator = hass.data[DOMAIN]["uuid"][
         "coordinator"
     ]
@@ -194,11 +194,10 @@ async def test_stop_job(hass: HomeAssistant) -> None:
         assert len(stop_command.mock_calls) == 0
 
 
-@freeze_time("2023-01-01 00:00")
+@pytest.mark.freeze_time("2023-01-01 00:00")
+@pytest.mark.usefixtures("init_integration")
 async def test_shutdown_system(hass: HomeAssistant) -> None:
     """Test the shutdown system button."""
-    await init_integration(hass, BUTTON_DOMAIN)
-
     entity_id = "button.octoprint_shutdown_system"
 
     # Test shutting down the system
@@ -219,11 +218,10 @@ async def test_shutdown_system(hass: HomeAssistant) -> None:
         assert state.state == "2023-01-01T00:00:00+00:00"
 
 
-@freeze_time("2023-01-01 00:00")
+@pytest.mark.freeze_time("2023-01-01 00:00")
+@pytest.mark.usefixtures("init_integration")
 async def test_reboot_system(hass: HomeAssistant) -> None:
     """Test the reboot system button."""
-    await init_integration(hass, BUTTON_DOMAIN)
-
     entity_id = "button.octoprint_reboot_system"
 
     # Test rebooting the system
@@ -246,11 +244,10 @@ async def test_reboot_system(hass: HomeAssistant) -> None:
         assert state.state == "2023-01-01T00:00:00+00:00"
 
 
-@freeze_time("2023-01-01 00:00")
+@pytest.mark.freeze_time("2023-01-01 00:00")
+@pytest.mark.usefixtures("init_integration")
 async def test_restart_octoprint(hass: HomeAssistant) -> None:
     """Test the restart octoprint button."""
-    await init_integration(hass, BUTTON_DOMAIN)
-
     entity_id = "button.octoprint_restart_octoprint"
 
     # Test restarting octoprint
