+++ tests/components/octoprint/test_number.py
@@ -1,9 +1,11 @@
 """The tests for OctoPrint number module."""
 
 from datetime import UTC, datetime
+from typing import Any
 from unittest.mock import patch
 
 from freezegun.api import FrozenDateTimeFactory
+import pytest
 from syrupy.assertion import SnapshotAssertion
 
 from homeassistant.components.number import (
@@ -15,55 +17,70 @@ from homeassistant.const import ATTR_ENTITY_ID, STATE_UNKNOWN
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers import entity_registry as er
 
-from . import init_integration
-
-from tests.common import snapshot_platform
+from tests.common import MockConfigEntry, snapshot_platform
 
 
+@pytest.fixture
+def platform() -> str:
+    """Fixture to specify platform."""
+    return NUMBER_DOMAIN
+
+
+@pytest.fixture
+def job() -> dict[str, Any]:
+    """Job fixture."""
+    return __standard_job()
+
+
+@pytest.mark.parametrize(
+    "printer",
+    [
+        {
+            "state": {
+                "flags": {"printing": True},
+                "text": "Operational",
+            },
+            "temperature": {
+                "tool0": {"actual": 18.83136, "target": 37.83136},
+                "tool1": {"actual": 21.0, "target": 31.0},
+                "bed": {"actual": 25.5, "target": 60.0},
+            },
+        },
+    ],
+)
+@pytest.mark.freeze_time(datetime(2020, 2, 20, 9, 10, 13, 543, tzinfo=UTC))
 async def test_numbers(
     hass: HomeAssistant,
-    freezer: FrozenDateTimeFactory,
     entity_registry: er.EntityRegistry,
     snapshot: SnapshotAssertion,
+    init_integration: MockConfigEntry,
 ) -> None:
     """Test the underlying number entities."""
-    printer = {
-        "state": {
-            "flags": {"printing": True},
-            "text": "Operational",
+    await snapshot_platform(hass, entity_registry, snapshot, init_integration.entry_id)
+
+
+@pytest.mark.parametrize(
+    "printer",
+    [
+        {
+            "state": {
+                "flags": {"printing": True, "paused": False},
+                "text": "Operational",
+            },
+            "temperature": {
+                "tool0": {"actual": 18.83136, "target": None},
+                "bed": {"actual": 25.5, "target": None},
+            },
         },
-        "temperature": {
-            "tool0": {"actual": 18.83136, "target": 37.83136},
-            "tool1": {"actual": 21.0, "target": 31.0},
-            "bed": {"actual": 25.5, "target": 60.0},
-        },
-    }
-    job = __standard_job()
-    freezer.move_to(datetime(2020, 2, 20, 9, 10, 13, 543, tzinfo=UTC))
-    config_entry = await init_integration(hass, "number", printer=printer, job=job)
-
-    await snapshot_platform(hass, entity_registry, snapshot, config_entry.entry_id)
-
-
+    ],
+)
+@pytest.mark.freeze_time(datetime(2020, 2, 20, 9, 10, 0))
+@pytest.mark.usefixtures("init_integration")
 async def test_numbers_no_target_temp(
     hass: HomeAssistant,
-    freezer: FrozenDateTimeFactory,
     entity_registry: er.EntityRegistry,
 ) -> None:
     """Test the number entities when target temperature is None."""
-    printer = {
-        "state": {
-            "flags": {"printing": True, "paused": False},
-            "text": "Operational",
-        },
-        "temperature": {
-            "tool0": {"actual": 18.83136, "target": None},
-            "bed": {"actual": 25.5, "target": None},
-        },
-    }
-    freezer.move_to(datetime(2020, 2, 20, 9, 10, 0))
-    await init_integration(hass, "number", printer=printer)
-
     state = hass.states.get("number.octoprint_extruder_temperature")
     assert state is not None
     assert state.state == STATE_UNKNOWN
@@ -79,22 +96,24 @@ async def test_numbers_no_target_temp(
     assert entry.unique_id == "uuid_bed_temperature"
 
 
+@pytest.mark.parametrize(
+    "printer",
+    [
+        {
+            "state": {
+                "flags": {"printing": False},
+                "text": "Operational",
+            },
+            "temperature": {"tool0": {"actual": 18.83136, "target": 25.0}},
+        },
+    ],
+)
+@pytest.mark.freeze_time(datetime(2020, 2, 20, 9, 10, 0))
+@pytest.mark.usefixtures("init_integration")
 async def test_set_tool_temp(
     hass: HomeAssistant,
-    freezer: FrozenDateTimeFactory,
 ) -> None:
     """Test setting tool temperature via number entity."""
-    printer = {
-        "state": {
-            "flags": {"printing": False},
-            "text": "Operational",
-        },
-        "temperature": {"tool0": {"actual": 18.83136, "target": 25.0}},
-    }
-    job = __standard_job()
-    freezer.move_to(datetime(2020, 2, 20, 9, 10, 0))
-    await init_integration(hass, "number", printer=printer, job=job)
-
     with patch(
         "pyoctoprintapi.OctoprintClient.set_tool_temperature"
     ) as mock_set_tool_temp:
@@ -114,22 +133,24 @@ async def test_set_tool_temp(
         mock_set_tool_temp.assert_called_with("tool0", 200)
 
 
+@pytest.mark.parametrize(
+    "printer",
+    [
+        {
+            "state": {
+                "flags": {"printing": False},
+                "text": "Operational",
+            },
+            "temperature": {"bed": {"actual": 20.0, "target": 50.0}},
+        },
+    ],
+)
+@pytest.mark.freeze_time(datetime(2020, 2, 20, 9, 10, 0))
+@pytest.mark.usefixtures("init_integration")
 async def test_set_bed_temp(
     hass: HomeAssistant,
-    freezer: FrozenDateTimeFactory,
 ) -> None:
     """Test setting bed temperature via number entity."""
-    printer = {
-        "state": {
-            "flags": {"printing": False},
-            "text": "Operational",
-        },
-        "temperature": {"bed": {"actual": 20.0, "target": 50.0}},
-    }
-    job = __standard_job()
-    freezer.move_to(datetime(2020, 2, 20, 9, 10, 0))
-    await init_integration(hass, "number", printer=printer, job=job)
-
     with patch(
         "pyoctoprintapi.OctoprintClient.set_bed_temperature"
     ) as mock_set_bed_temp:
@@ -149,25 +170,27 @@ async def test_set_bed_temp(
         mock_set_bed_temp.assert_called_with(80)
 
 
+@pytest.mark.parametrize(
+    "printer",
+    [
+        {
+            "state": {
+                "flags": {"printing": False},
+                "text": "Operational",
+            },
+            "temperature": {
+                "tool0": {"actual": 20.0, "target": 30.0},
+                "tool1": {"actual": 21.0, "target": 31.0},
+            },
+        },
+    ],
+)
+@pytest.mark.freeze_time(datetime(2020, 2, 20, 9, 10, 0))
+@pytest.mark.usefixtures("init_integration")
 async def test_set_tool_n_temp(
     hass: HomeAssistant,
-    freezer: FrozenDateTimeFactory,
 ) -> None:
     """Test setting tool temperature via number entity when multiple tools are present."""
-    printer = {
-        "state": {
-            "flags": {"printing": False},
-            "text": "Operational",
-        },
-        "temperature": {
-            "tool0": {"actual": 20.0, "target": 30.0},
-            "tool1": {"actual": 21.0, "target": 31.0},
-        },
-    }
-    job = __standard_job()
-    freezer.move_to(datetime(2020, 2, 20, 9, 10, 0))
-    await init_integration(hass, "number", printer=printer, job=job)
-
     with patch(
         "pyoctoprintapi.OctoprintClient.set_tool_temperature"
     ) as mock_set_tool_temp:
@@ -187,16 +210,14 @@ async def test_set_tool_n_temp(
         mock_set_tool_temp.assert_called_with("tool1", 41)
 
 
+@pytest.mark.parametrize("printer", [None])
+@pytest.mark.freeze_time(datetime(2020, 2, 20, 9, 10, 0))
+@pytest.mark.usefixtures("init_integration")
 async def test_numbers_printer_disconnected(
     hass: HomeAssistant,
     freezer: FrozenDateTimeFactory,
-    entity_registry: er.EntityRegistry,
 ) -> None:
     """Test number entities when printer is disconnected."""
-    job = __standard_job()
-    freezer.move_to(datetime(2020, 2, 20, 9, 10, 0))
-    await init_integration(hass, "number", printer=None, job=job)
-
     # When printer is disconnected, no number entities should be created
     state = hass.states.get("number.octoprint_tool0_temperature")
     assert state is None
