+++ tests/components/octoprint/test_sensor.py
@@ -1,33 +1,49 @@
 """The tests for Octoptint binary sensor module."""
 
 from datetime import UTC, datetime
+from typing import Any
 
 from freezegun.api import FrozenDateTimeFactory
+import pytest
 
 from homeassistant.const import STATE_UNAVAILABLE, STATE_UNKNOWN, UnitOfInformation
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers import entity_registry as er
 
-from . import init_integration
+from . import DEFAULT_JOB
 
 
+@pytest.fixture
+def platform() -> str:
+    """Fixture to specify platform."""
+    return "sensor"
+
+
+@pytest.fixture
+def job() -> dict[str, Any]:
+    """Job fixture."""
+    return __standard_job()
+
+
+@pytest.mark.parametrize(
+    "printer",
+    [
+        {
+            "state": {
+                "flags": {"printing": True},
+                "text": "Operational",
+            },
+            "temperature": {"tool1": {"actual": 18.83136, "target": 37.83136}},
+        },
+    ],
+)
+@pytest.mark.freeze_time(datetime(2020, 2, 20, 9, 10, 13, 543, tzinfo=UTC))
+@pytest.mark.usefixtures("init_integration")
 async def test_sensors(
     hass: HomeAssistant,
-    freezer: FrozenDateTimeFactory,
     entity_registry: er.EntityRegistry,
 ) -> None:
     """Test the underlying sensors."""
-    printer = {
-        "state": {
-            "flags": {"printing": True},
-            "text": "Operational",
-        },
-        "temperature": {"tool1": {"actual": 18.83136, "target": 37.83136}},
-    }
-    job = __standard_job()
-    freezer.move_to(datetime(2020, 2, 20, 9, 10, 13, 543, tzinfo=UTC))
-    await init_integration(hass, "sensor", printer=printer, job=job)
-
     state = hass.states.get("sensor.octoprint_job_percentage")
     assert state is not None
     assert state.state == "50"
@@ -93,22 +109,27 @@ async def test_sensors(
     assert entry.unique_id == "Current File Size-uuid"
 
 
+@pytest.mark.parametrize("job", [DEFAULT_JOB])
+@pytest.mark.parametrize(
+    "printer",
+    [
+        {
+            "state": {
+                "flags": {"printing": True, "paused": False},
+                "text": "Operational",
+            },
+            "temperature": {"tool1": {"actual": 18.83136, "target": None}},
+        },
+    ],
+)
+@pytest.mark.freeze_time(datetime(2020, 2, 20, 9, 10, 0))
+@pytest.mark.usefixtures("init_integration")
 async def test_sensors_no_target_temp(
     hass: HomeAssistant,
     freezer: FrozenDateTimeFactory,
     entity_registry: er.EntityRegistry,
 ) -> None:
     """Test the underlying sensors."""
-    printer = {
-        "state": {
-            "flags": {"printing": True, "paused": False},
-            "text": "Operational",
-        },
-        "temperature": {"tool1": {"actual": 18.83136, "target": None}},
-    }
-    freezer.move_to(datetime(2020, 2, 20, 9, 10, 0))
-    await init_integration(hass, "sensor", printer=printer)
-
     state = hass.states.get("sensor.octoprint_actual_tool1_temp")
     assert state is not None
     assert state.state == "18.83"
@@ -138,23 +159,26 @@ async def test_sensors_no_target_temp(
     assert entry.unique_id == "Current File Size-uuid"
 
 
+@pytest.mark.parametrize(
+    "printer",
+    [
+        {
+            "state": {
+                "flags": {"printing": False},
+                "text": "Operational",
+            },
+            "temperature": {"tool1": {"actual": 18.83136, "target": None}},
+        },
+    ],
+)
+@pytest.mark.freeze_time(datetime(2020, 2, 20, 9, 10, 0))
+@pytest.mark.usefixtures("init_integration")
 async def test_sensors_paused(
     hass: HomeAssistant,
     freezer: FrozenDateTimeFactory,
     entity_registry: er.EntityRegistry,
 ) -> None:
     """Test the underlying sensors."""
-    printer = {
-        "state": {
-            "flags": {"printing": False},
-            "text": "Operational",
-        },
-        "temperature": {"tool1": {"actual": 18.83136, "target": None}},
-    }
-    job = __standard_job()
-    freezer.move_to(datetime(2020, 2, 20, 9, 10, 0))
-    await init_integration(hass, "sensor", printer=printer, job=job)
-
     state = hass.states.get("sensor.octoprint_start_time")
     assert state is not None
     assert state.state == STATE_UNKNOWN
@@ -170,16 +194,15 @@ async def test_sensors_paused(
     assert entry.unique_id == "Estimated Finish Time-uuid"
 
 
+@pytest.mark.parametrize("printer", [None])
+@pytest.mark.freeze_time(datetime(2020, 2, 20, 9, 10, 0))
+@pytest.mark.usefixtures("init_integration")
 async def test_sensors_printer_disconnected(
     hass: HomeAssistant,
     freezer: FrozenDateTimeFactory,
     entity_registry: er.EntityRegistry,
 ) -> None:
     """Test the underlying sensors."""
-    job = __standard_job()
-    freezer.move_to(datetime(2020, 2, 20, 9, 10, 0))
-    await init_integration(hass, "sensor", printer=None, job=job)
-
     state = hass.states.get("sensor.octoprint_job_percentage")
     assert state is not None
     assert state.state == "50"
