+++ tests/components/octoprint/test_services.py
@@ -2,6 +2,8 @@
 
 from unittest.mock import patch
 
+import pytest
+
 from homeassistant.components.octoprint.const import (
     CONF_BAUDRATE,
     DOMAIN,
@@ -11,15 +13,18 @@ from homeassistant.const import ATTR_DEVICE_ID, CONF_PORT, CONF_PROFILE_NAME
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers import device_registry as dr
 
-from . import init_integration
+
+@pytest.fixture
+def platform() -> str:
+    """Fixture to specify platform."""
+    return "sensor"
 
 
+@pytest.mark.usefixtures("init_integration")
 async def test_connect_default(
     hass: HomeAssistant, device_registry: dr.DeviceRegistry
 ) -> None:
     """Test the connect to printer service."""
-    await init_integration(hass, "sensor")
-
     device = dr.async_entries_for_config_entry(device_registry, "uuid")[0]
 
     # Test pausing the printer when it is printing
@@ -39,12 +44,11 @@ async def test_connect_default(
         )
 
 
+@pytest.mark.usefixtures("init_integration")
 async def test_connect_all_arguments(
     hass: HomeAssistant, device_registry: dr.DeviceRegistry
 ) -> None:
     """Test the connect to printer service."""
-    await init_integration(hass, "sensor")
-
     device = dr.async_entries_for_config_entry(device_registry, "uuid")[0]
 
     # Test pausing the printer when it is printing
